+++ src/entity/bookmark/bookmark.args.ts
@@ -1,5 +1,5 @@
 import { BookmarkOrderFields } from '../../types/enums.js';
-import { ArgsType, Field, Float, ID, InputType } from 'type-graphql';
+import { ArgsType, Field, ID, InputType, Int } from 'type-graphql';
 import { OrderByArgs, PaginatedFilterArgs } from '../base/base.args.js';
 import { examples } from '../../modules/engine/rest/example.consts.js';
 import { ObjParamsType } from '../../modules/rest/decorators/ObjParamsType.js';
@@ -46,7 +46,7 @@ export class BookmarkFilterArgs {
 	@ObjField(() => [String], { nullable: true, description: 'filter by Episode Ids', isID: true })
 	episodeIDs?: Array<string>;
 
-	@Field(() => Float, { nullable: true })
+	@Field(() => Int, { nullable: true })
 	@ObjField({ nullable: true, description: 'filter by Creation timestamp', min: 0, example: examples.timestamp })
 	since?: number;
