+++ src/entity/folder/folder.args.ts
@@ -1,5 +1,5 @@
 import { AlbumType, FolderOrderFields, FolderType, ListType } from '../../types/enums.js';
-import { ArgsType, Field, Float, ID, InputType, Int } from 'type-graphql';
+import { ArgsType, Field, ID, InputType, Int } from 'type-graphql';
 import { FilterArgs, OrderByArgs, PaginatedFilterArgs } from '../base/base.args.js';
 import { examples } from '../../modules/engine/rest/example.consts.js';
 import { ObjParamsType } from '../../modules/rest/decorators/ObjParamsType.js';
@@ -134,7 +134,7 @@ export class FolderFilterArgs {
 	@ObjField(() => [String], { nullable: true, description: 'filter by Folder Ids', isID: true })
 	ids?: Array<string>;
 
-	@Field(() => Float, { nullable: true })
+	@Field(() => Int, { nullable: true })
 	@ObjField({ nullable: true, description: 'filter by Creation timestamp', min: 0, example: examples.timestamp })
 	since?: number;
