+++ src/entity/nowplaying/nowplaying.ts
@@ -1,11 +1,11 @@
 import { Track, TrackQL } from '../track/track.js';
 import { Episode, EpisodeQL } from '../episode/episode.js';
 import { User } from '../user/user.js';
-import { Field, Float, ID, ObjectType } from 'type-graphql';
+import { Field, ID, Int, ObjectType } from 'type-graphql';
 
 @ObjectType()
 export class NowPlaying {
-	@Field(() => Float)
+	@Field(() => Int)
 	time!: number;
 
 	@Field(() => TrackQL, { nullable: true })
