+++ src/entity/playlist/playlist.ts
@@ -1,6 +1,6 @@
 import { User } from '../user/user.js';
 import { PlaylistEntry, PlaylistEntryQL } from '../playlistentry/playlist-entry.js';
-import { Field, Float, ID, Int, ObjectType } from 'type-graphql';
+import { Field, ID, Int, ObjectType } from 'type-graphql';
 import { Collection, Entity, ManyToOne, OneToMany, ORM_INT, Property, Reference } from '../../modules/orm/index.js';
 import { Base, Index, IndexGroup, PaginatedResponse } from '../base/base.js';
 import { State, StateQL } from '../state/state.js';
@@ -27,7 +27,7 @@ export class Playlist extends Base {
 	@Property(() => Boolean)
 	isPublic: boolean = false;
 
-	@Field(() => Float)
+	@Field(() => Int)
 	@Property(() => ORM_INT)
 	duration: number = 0;
